إصلاح: تصحيح import path وإصلاح type casting
المشاكل المحلولة:
❌ import خاطئ: '../../services/providers.dart'
✅ import صحيح: '../../providers/repository_providers.dart'

❌ type error: clamp() returns num
✅ type fix: .clamp().toDouble()

هذا يحل أخطاء البناء في GitHub Actions:
- roomsRepoProvider لم يكن موجوداً
- paymentsRepoProvider لم يكن موجوداً
- bookingsRepoProvider لم يكن موجوداً
- whatsappServiceProvider لم يكن موجوداً
- خطأ type casting في newRemaining

+++ mobile/lib/screens/payments/booking_payment_screen.dart
@@ -8,7 +8,7 @@ import '../../components/app_scaffold.dart';
 import '../../services/local_db.dart' as db;
 import '../../models/payment_models.dart';
 import '../../components/widgets/payment_widgets.dart';
-import '../../services/providers.dart';
+import '../../providers/repository_providers.dart';
 import '../../utils/time.dart';
 import 'payment_history_screen.dart';
 import '../../mixins/sync_on_exit_mixin.dart';
@@ -1116,7 +1116,7 @@ class _BookingPaymentScreenState extends ConsumerState<BookingPaymentScreen>
     final currentTotal = currentNights * roomRate;
     final allPayments = await paymentsRepo.paymentsByBooking(widget.booking.id).first;
     final totalPaid = allPayments.fold<double>(0, (s, p) => s + p.amount);
-    final newRemaining = (currentTotal - totalPaid).clamp(0.0, currentTotal);
+    final newRemaining = (currentTotal - totalPaid).clamp(0.0, currentTotal).toDouble();
 
     // إرسال رسالة واتساب
     final cleanedPhone = _cleanAndFormatPhone(_currentGuestPhone);
